Part 4 - ex 4.21

+++ part3/blog_backend/controllers/blogs.js
@@ -68,9 +68,20 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete(`/:id`, async (request, response, next) => {
-  await Blog.findByIdAndDelete(request.params.id)
+  const blog = await Blog.findById(request.params.id)
 
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  const user = await User.findById(decodedToken.id)
+
+  if ( blog.user.toString() === user.id.toString() ) {
+    await Blog.findByIdAndDelete(request.params.id)
+  
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'invalid token' })
+  }
+
 })
 
 module.exports = blogsRouter
